Make upload_to_tpa primary action and tpa an alias
